\name{distfree.cr-package}
\alias{distfree.cr-package}
\docType{package}
\title{
Distribution-free confidence region (distfree.cr)
}
\description{
The disfree.cr package was developed to implement a novel geometry-based method introduced by Zhiqiu Hu and Rong-cai Yang for constructing confidence regions without the need to know the sampling distribution of estimated parameters for two or more variables.
}
\details{
\tabular{ll}{
Package: \tab distfree.cr\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-11-23\cr
License: \tab GPL (>2.0)\cr
}

}
\author{
Zhiqiu Hu and Rong-cai Yang

Maintainer: Zhiqiu Hu <zhiqiu.hu@gmail.com>
}
\keyword{ package }
