\name{plot.distfree.cr}
\alias{plot.distfree.cr}
\title{
plot.distfree.cr
}
\description{
Plot an object that is returned by the distfree.cr function.
}
\usage{
\method{plot}{distfree.cr}(x, show.points = T, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{An object returned by the distfree.cr function.}
  \item{show.points}{A logical indicator of whether or not the original data are plotted.}
     \item{\dots}{Other parameters that can be passed to the \code{plot} function.}
}

